LayoutTests/http/wpt/wk-web-animations/timing-model/animation-effect-timing.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Check that the AnimationEffectTiming interface is defined and does not expose a constructor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

internals.pauseTimeline(document.timeline);

test(t => {
    assert_own_property(window, "AnimationEffectTiming");
}, "The AnimationEffectTiming interface is defined.");

test(t => {
    assert_throws(new TypeError, () => new AnimationEffectTiming);
}, "The AnimationEffectTiming interface cannot be constructed.");

</script>
</body>
